Started refactoring code into model.py
